1.1.x

+++ src/commands/poll.js
@@ -2,30 +2,27 @@ module.exports = {
     name: 'poll',
     desc: 'For Bot Owners/Admins and President to create polls with an upvote and downvote option.',
     usage: ';poll (poll message) (upvote option) (downvote option)',
-    examples: ";poll \"Should we separate the Speaker of the House " + 
-    "into one for House and one for Senate?\" " + 
-    '"Separate it!" "Don\'t separate the role" // Creates a poll with the content shown.',
+    examples: ";poll \"Should we separate the Speaker of the House " +
+        "into one for House and one for Senate?\" " +
+        '"Separate it!" "Don\'t separate the role" // Creates a poll with the content shown.',
     execute(msg, bot, args) {
         const Discord = require("discord.js");
-        if (!msg.member.roles.find(r => r.name === "Bot Admin") && 
-            !msg.member.roles.find(r => r.name === "Bot Owner") && 
-            !msg.member.roles.find(r => r.name === "President")) 
-        return msg.channel.send("You lack permissions to use this command.");
+        if (!msg.member.roles.find(r => r.name === "Bot Admin") &&
+            !msg.member.roles.find(r => r.name === "Bot Owner") &&
+            !msg.member.roles.find(r => r.name === "President"))
+            return msg.channel.send("You lack permissions to use this command.");
         msg.channel.send("Creating poll...")
             .then(m => {
                 msg.delete();
                 let argString = args.join(" "); // this turns the args into a string...
                 let parsedArgs = argString.split('"'); // ...then splits it by the " keychar... 
-                console.log(parsedArgs); // ...and then...
-                parsedArgs.splice(0, 1) // ...removes the blank value in front...
-                    .splice(1, 1) // ...removes the blank value after the first argument...
-                    .splice(2, 1) // ...removes the blank value after the second argument...
-                    .splice(3, 1); // ...and removes the blank value at the end!
-                console.log(parsedArgs); // figured this out myself, pretty proud of it tbh
+                for (let i = 0; i < parsedArgs.length; i++) {
+                    if (parsedArgs[i] === "" || parsedArgs[i] === " ") parsedArgs.splice(i, 1);
+                }
                 const upvote = msg.guild.emojis.find(e => e.name === "Upvote");
                 const downvote = msg.guild.emojis.find(e => e.name === "Downvote");
                 msg.channel.send(`@everyone, ${parsedArgs[0]} Vote here:\n\n` +
-                                 `${upvote}: ${parsedArgs[1]}\n\n${downvote}: ${parsedArgs[2]}`)
+                        `${upvote}: ${parsedArgs[1]}\n\n${downvote}: ${parsedArgs[2]}`)
                     .then(async msg => { // required to get awaits to work
                         m.delete();
                         msg.channel.send("Reacting...")
